+++ src/features/Status/StatusPage.tsx
@@ -10,7 +10,6 @@ import {
   LinkDuo,
   Paragraph,
 } from '../../shared/Elements';
-// import Sidebar from '../Sidebar/Sidebar';
 
 import {
   FrontendRoute,
@@ -96,16 +95,34 @@ class StatusPage extends React.Component<IStatusPageProps, {}> {
                   </Paragraph>
                 </Flex>
               )}
+              ) : (
+              <Flex
+                flexDirection={'column'}
+                style={{ marginTop: '1em' }}
+                alignItems={'center'}
+              >
+                <Paragraph
+                  color={theme.colors.black80}
+                  textAlign={'center'}
+                  marginBottom={'3rem'}
+                >
+                  {CONSTANTS.NONE_STATUS_TEXT}
+                </Paragraph>
+                <LinkDuo to={FrontendRoute.CREATE_APPLICATION_PAGE}>
+                  <Button type="button">Apply</Button>
+                </LinkDuo>
+              </Flex>
+              ) }
+              <BackgroundImage
+                right={'0px'}
+                bottom={'0px'}
+                src={Background}
+                imgHeight={'87%'}
+              />
             </div>
           ) : (
             <ConfirmationEmailSentComponent />
           )}
-          <BackgroundImage
-            right={'0px'}
-            bottom={'0px'}
-            src={Background}
-            imgHeight={'87%'}
-          />
         </Box>
       </Flex>
     );